+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - A test message can now be sent to specific logger, from the loggers list.
 ### Changed
+- Status (in loggers list) have now their own column.
 - The `README.md` displays badges.
 ### Removed
 - The section "install from GitHub" has been removed from `README.md`.